Problem: unable to reproduce JAXB issue
Solution:
To "get there" in terms of reproducing JAXB issue, we first had to solve another error, which originated from `ConnectionTest.java`. See comment describing the rationale for this change under JDK11.

Another step on the way to fixing JAXB stuff, making it possible to compile the MAGE world under JDK11 was doing some things that are referred to in #8748, namely:

 - Add flake.nix with OpenJDK11 headless and JRE
 - Add .envrc for direnv-compatible development
 - Test that stuff works with VSCode java extensions (NB! for maximum awesomeness, install cab404's direnv extension, which will hook up into direnv)